check-in2:loging enabled custom error message JPA repository argument change adding checkin-2.txt file

+++ src/main/java/com/loiane/springcloudhello/repository/departmentrepository.java
@@ -7,6 +7,6 @@ import com.loiane.springcloudhello.model.Contact;
 import com.loiane.springcloudhello.model.department;
 
 @Repository
-public interface departmentrepository extends JpaRepository<department, Long>{
+public interface departmentrepository extends JpaRepository<department, Integer>{
 
 }
